requirements.txt
argparse~=1.4.0
numpy~=1.19.1
opencv-python~=4.5.1.48
PyYAML~=5.4.1
tqdm~=4.60.0

